+++ dubbo-config/dubbo-config-spring/src/main/java/com/alibaba/dubbo/config/spring/beans/factory/annotation/ServiceAnnotationBeanPostProcessor.java
@@ -1,0 +1,304 @@
+package com.alibaba.dubbo.config.spring.beans.factory.annotation;
+
+import com.alibaba.dubbo.common.logger.Logger;
+import com.alibaba.dubbo.common.logger.LoggerFactory;
+import com.alibaba.dubbo.config.annotation.Service;
+import com.alibaba.dubbo.config.spring.ServiceBean;
+import com.alibaba.dubbo.config.spring.context.annotation.DubboClassPathBeanDefinitionScanner;
+import org.springframework.beans.BeansException;
+import org.springframework.beans.factory.BeanClassLoaderAware;
+import org.springframework.beans.factory.config.BeanDefinition;
+import org.springframework.beans.factory.config.BeanDefinitionHolder;
+import org.springframework.beans.factory.config.ConfigurableListableBeanFactory;
+import org.springframework.beans.factory.config.RuntimeBeanReference;
+import org.springframework.beans.factory.support.*;
+import org.springframework.context.EnvironmentAware;
+import org.springframework.context.ResourceLoaderAware;
+import org.springframework.core.env.Environment;
+import org.springframework.core.io.ResourceLoader;
+import org.springframework.core.type.filter.AnnotationTypeFilter;
+import org.springframework.util.Assert;
+import org.springframework.util.ClassUtils;
+import org.springframework.util.ObjectUtils;
+import org.springframework.util.StringUtils;
+
+import java.util.*;
+
+import static org.springframework.beans.factory.support.BeanDefinitionBuilder.rootBeanDefinition;
+import static org.springframework.beans.factory.support.BeanDefinitionReaderUtils.registerWithGeneratedName;
+import static org.springframework.core.annotation.AnnotationUtils.findAnnotation;
+import static org.springframework.util.ClassUtils.resolveClassName;
+
+/**
+ * {@link Service} Annotation
+ * {@link BeanDefinitionRegistryPostProcessor Bean Definition Registry Post Processor}
+ *
+ * @author <a href="mailto:mercyblitz@gmail.com">Mercy</a>
+ * @since 2.5.8
+ */
+public class ServiceAnnotationBeanPostProcessor implements BeanDefinitionRegistryPostProcessor, EnvironmentAware,
+        ResourceLoaderAware, BeanClassLoaderAware {
+
+    private final Logger logger = LoggerFactory.getLogger(getClass());
+
+    private final Set<String> packagesToScan;
+
+    private Environment environment;
+
+    private ResourceLoader resourceLoader;
+
+    private ClassLoader classLoader;
+
+    public ServiceAnnotationBeanPostProcessor(String... packagesToScan) {
+        this(Arrays.asList(packagesToScan));
+    }
+
+    public ServiceAnnotationBeanPostProcessor(Collection<String> packagesToScan) {
+        this(new LinkedHashSet<String>(packagesToScan));
+    }
+
+    public ServiceAnnotationBeanPostProcessor(Set<String> packagesToScan) {
+        this.packagesToScan = packagesToScan;
+    }
+
+    @Override
+    public void postProcessBeanDefinitionRegistry(BeanDefinitionRegistry registry) throws BeansException {
+
+        Set<String> resolvedPackagesToScan = resolvePackagesToScan(packagesToScan);
+
+        registerServiceBeans(resolvedPackagesToScan, registry);
+
+    }
+
+
+    /**
+     * Registers Beans whose classes was annotated {@link Service}
+     *
+     * @param packagesToScan The base packages to scan
+     * @param registry       {@link BeanDefinitionRegistry}
+     */
+    private void registerServiceBeans(Set<String> packagesToScan, BeanDefinitionRegistry registry) {
+
+        DubboClassPathBeanDefinitionScanner dubboClassPathBeanDefinitionScanner =
+                new DubboClassPathBeanDefinitionScanner(registry, environment, resourceLoader);
+
+        dubboClassPathBeanDefinitionScanner.addIncludeFilter(new AnnotationTypeFilter(Service.class));
+
+        for (String packageToScan : packagesToScan) {
+
+            Set<BeanDefinitionHolder> beanDefinitionHolders = dubboClassPathBeanDefinitionScanner.doScan(packageToScan);
+
+            for (BeanDefinitionHolder beanDefinitionHolder : beanDefinitionHolders) {
+                registerServiceBean(beanDefinitionHolder, registry);
+            }
+
+            if (logger.isInfoEnabled()) {
+                logger.info(beanDefinitionHolders.size() + " annotated @Service Components { " +
+                        beanDefinitionHolders +
+                        " } were scanned under package[" + packageToScan + "]");
+            }
+        }
+
+    }
+
+    /**
+     * Registers {@link ServiceBean} from new annotated {@link Service} {@link BeanDefinition}
+     *
+     * @param beanDefinitionHolder
+     * @param registry
+     * @see ServiceBean
+     * @see BeanDefinition
+     */
+    private void registerServiceBean(BeanDefinitionHolder beanDefinitionHolder, BeanDefinitionRegistry registry) {
+
+        Class<?> beanClass = resolveClass(beanDefinitionHolder);
+
+        Service service = findAnnotation(beanClass, Service.class);
+
+        Class<?> interfaceClass = resolveServiceInterfaceClass(beanClass, service);
+
+        String beanName = beanDefinitionHolder.getBeanName();
+
+        AbstractBeanDefinition serviceBeanDefinition = buildServiceBeanDefinition(service, interfaceClass, beanName);
+
+        registerWithGeneratedName(serviceBeanDefinition, registry);
+
+    }
+
+    private Class<?> resolveServiceInterfaceClass(Class<?> annotatedServiceBeanClass, Service service) {
+
+        Class<?> interfaceClass = service.interfaceClass();
+
+        if (void.class.equals(interfaceClass)) {
+
+            interfaceClass = null;
+
+            String interfaceClassName = service.interfaceName();
+
+            if (StringUtils.hasText(interfaceClassName)) {
+                if (ClassUtils.isPresent(interfaceClassName, classLoader)) {
+                    interfaceClass = resolveClassName(interfaceClassName, classLoader);
+                }
+            }
+
+        }
+
+        if (interfaceClass == null) {
+
+            Class<?>[] allInterfaces = annotatedServiceBeanClass.getInterfaces();
+
+            if (allInterfaces.length > 0) {
+                interfaceClass = allInterfaces[0];
+            }
+
+        }
+
+        Assert.notNull(interfaceClass,
+                "@Service interfaceClass() or interfaceName() or interface class must be present!");
+
+        Assert.isTrue(interfaceClass.isInterface(),
+                "The type that was annotated @Service is not an interface!");
+
+        return interfaceClass;
+    }
+
+    private Class<?> resolveClass(BeanDefinitionHolder beanDefinitionHolder) {
+
+        BeanDefinition beanDefinition = beanDefinitionHolder.getBeanDefinition();
+
+        return resolveClass(beanDefinition);
+
+    }
+
+    private Class<?> resolveClass(BeanDefinition beanDefinition) {
+
+        String beanClassName = beanDefinition.getBeanClassName();
+
+        return resolveClassName(beanClassName, classLoader);
+
+    }
+
+    private Set<String> resolvePackagesToScan(Set<String> packagesToScan) {
+        Set<String> resolvedPackagesToScan = new LinkedHashSet<String>(packagesToScan.size());
+        for (String packageToScan : packagesToScan) {
+            String resolvedPackageToScan = environment.resolvePlaceholders(packageToScan);
+            resolvedPackagesToScan.add(resolvedPackageToScan);
+        }
+        return resolvedPackagesToScan;
+    }
+
+    private AbstractBeanDefinition buildServiceBeanDefinition(Service service, Class<?> interfaceClass,
+                                                              String annotatedServiceBeanName) {
+
+        BeanDefinitionBuilder builder = rootBeanDefinition(ServiceBean.class)
+                .addConstructorArgValue(service)
+                // References "ref" property to annotated-@Service Bean
+                .addPropertyReference("ref", annotatedServiceBeanName)
+                .addPropertyValue("interfaceClass", interfaceClass);
+
+        /**
+         * Add {@link com.alibaba.dubbo.config.ProviderConfig} Bean reference
+         */
+        String providerConfigBeanName = service.provider();
+        if (StringUtils.hasText(providerConfigBeanName)) {
+            addPropertyReference(builder, "provider", providerConfigBeanName);
+        }
+
+        /**
+         * Add {@link com.alibaba.dubbo.config.MonitorConfig} Bean reference
+         */
+        String monitorConfigBeanName = service.monitor();
+        if (StringUtils.hasText(monitorConfigBeanName)) {
+            addPropertyReference(builder, "monitor", monitorConfigBeanName);
+        }
+
+        /**
+         * Add {@link com.alibaba.dubbo.config.ApplicationConfig} Bean reference
+         */
+        String applicationConfigBeanName = service.application();
+        if (StringUtils.hasText(applicationConfigBeanName)) {
+            addPropertyReference(builder, "application", applicationConfigBeanName);
+        }
+
+        /**
+         * Add {@link com.alibaba.dubbo.config.ModuleConfig} Bean reference
+         */
+        String moduleConfigBeanName = service.module();
+        if (StringUtils.hasText(moduleConfigBeanName)) {
+            addPropertyReference(builder, "module", moduleConfigBeanName);
+        }
+
+
+        /**
+         * Add {@link com.alibaba.dubbo.config.RegistryConfig} Bean reference
+         */
+        String[] registryConfigBeanNames = service.registry();
+
+        List<RuntimeBeanReference> registryRuntimeBeanReferences = toRuntimeBeanReferences(registryConfigBeanNames);
+
+        if (!registryRuntimeBeanReferences.isEmpty()) {
+            builder.addPropertyValue("registries", registryRuntimeBeanReferences);
+        }
+
+        /**
+         * Add {@link com.alibaba.dubbo.config.ProtocolConfig} Bean reference
+         */
+        String[] protocolConfigBeanNames = service.protocol();
+
+        List<RuntimeBeanReference> protocolRuntimeBeanReferences = toRuntimeBeanReferences(protocolConfigBeanNames);
+
+        if (!registryRuntimeBeanReferences.isEmpty()) {
+            builder.addPropertyValue("protocols", protocolRuntimeBeanReferences);
+        }
+
+        return builder.getBeanDefinition();
+
+    }
+
+
+    private ManagedList<RuntimeBeanReference> toRuntimeBeanReferences(String... beanNames) {
+
+        ManagedList<RuntimeBeanReference> runtimeBeanReferences = new ManagedList<RuntimeBeanReference>();
+
+        if (!ObjectUtils.isEmpty(beanNames)) {
+
+            for (String beanName : beanNames) {
+
+                String resolvedBeanName = environment.resolvePlaceholders(beanName);
+
+                runtimeBeanReferences.add(new RuntimeBeanReference(resolvedBeanName));
+            }
+
+        }
+
+        return runtimeBeanReferences;
+
+    }
+
+    private void addPropertyReference(BeanDefinitionBuilder builder, String propertyName, String beanName) {
+        String resolvedBeanName = environment.resolvePlaceholders(beanName);
+        builder.addPropertyReference(propertyName, resolvedBeanName);
+    }
+
+
+    @Override
+    public void postProcessBeanFactory(ConfigurableListableBeanFactory beanFactory) throws BeansException {
+
+    }
+
+    @Override
+    public void setEnvironment(Environment environment) {
+        this.environment = environment;
+    }
+
+    @Override
+    public void setResourceLoader(ResourceLoader resourceLoader) {
+        this.resourceLoader = resourceLoader;
+    }
+
+    @Override
+    public void setBeanClassLoader(ClassLoader classLoader) {
+        this.classLoader = classLoader;
+    }
+
+}
