+++ dubbo-config/dubbo-config-spring/src/main/java/com/alibaba/dubbo/config/spring/context/annotation/DubboComponentScanRegistrar.java
@@ -1,39 +1,26 @@
 package com.alibaba.dubbo.config.spring.context.annotation;
 
 import com.alibaba.dubbo.config.annotation.Service;
-import com.alibaba.dubbo.config.spring.ServiceBean;
 import com.alibaba.dubbo.config.spring.beans.factory.annotation.ReferenceAnnotationBeanPostProcessor;
+import com.alibaba.dubbo.config.spring.beans.factory.annotation.ServiceAnnotationBeanPostProcessor;
 import com.alibaba.dubbo.config.spring.util.BeanRegistrar;
-import org.apache.commons.logging.Log;
-import org.apache.commons.logging.LogFactory;
-import org.springframework.beans.factory.BeanClassLoaderAware;
 import org.springframework.beans.factory.BeanFactory;
 import org.springframework.beans.factory.config.BeanDefinition;
-import org.springframework.beans.factory.config.BeanDefinitionHolder;
-import org.springframework.beans.factory.config.RuntimeBeanReference;
 import org.springframework.beans.factory.support.AbstractBeanDefinition;
 import org.springframework.beans.factory.support.BeanDefinitionBuilder;
+import org.springframework.beans.factory.support.BeanDefinitionReaderUtils;
 import org.springframework.beans.factory.support.BeanDefinitionRegistry;
-import org.springframework.beans.factory.support.ManagedList;
-import org.springframework.context.EnvironmentAware;
-import org.springframework.context.ResourceLoaderAware;
 import org.springframework.context.annotation.ImportBeanDefinitionRegistrar;
 import org.springframework.core.annotation.AnnotationAttributes;
-import org.springframework.core.env.Environment;
-import org.springframework.core.io.ResourceLoader;
 import org.springframework.core.type.AnnotationMetadata;
-import org.springframework.core.type.filter.AnnotationTypeFilter;
-import org.springframework.util.Assert;
 import org.springframework.util.ClassUtils;
-import org.springframework.util.ObjectUtils;
-import org.springframework.util.StringUtils;
 
-import java.util.*;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.LinkedHashSet;
+import java.util.Set;
 
 import static org.springframework.beans.factory.support.BeanDefinitionBuilder.rootBeanDefinition;
-import static org.springframework.beans.factory.support.BeanDefinitionReaderUtils.registerWithGeneratedName;
-import static org.springframework.core.annotation.AnnotationUtils.findAnnotation;
-import static org.springframework.util.ClassUtils.resolveClassName;
 
 /**
  * Dubbo {@link DubboComponentScan} Bean Registrar
@@ -42,228 +29,38 @@ import static org.springframework.util.ClassUtils.resolveClassName;
  * @see Service
  * @see DubboComponentScan
  * @see ImportBeanDefinitionRegistrar
+ * @see ServiceAnnotationBeanPostProcessor
+ * @see ReferenceAnnotationBeanPostProcessor
  * @since 2.5.7
  */
-public class DubboComponentScanRegistrar implements ImportBeanDefinitionRegistrar, ResourceLoaderAware,
-        EnvironmentAware, BeanClassLoaderAware {
-
-    private final Log logger = LogFactory.getLog(getClass());
-
-    private ResourceLoader resourceLoader;
-
-    private Environment environment;
-
-    private ClassLoader classLoader;
+public class DubboComponentScanRegistrar implements ImportBeanDefinitionRegistrar {
 
     @Override
     public void registerBeanDefinitions(AnnotationMetadata importingClassMetadata, BeanDefinitionRegistry registry) {
 
         Set<String> packagesToScan = getPackagesToScan(importingClassMetadata);
 
-        registerServiceBeans(packagesToScan, registry);
+        registerServiceAnnotationBeanPostProcessor(packagesToScan, registry);
 
         registerReferenceAnnotationBeanPostProcessor(registry);
 
     }
 
     /**
-     * Registers Beans whose classes was annotated {@link Service}
+     * Registers {@link ServiceAnnotationBeanPostProcessor}
      *
-     * @param packagesToScan The base packages to scan
+     * @param packagesToScan packages to scan without resolving placeholders
      * @param registry       {@link BeanDefinitionRegistry}
+     * @since 2.5.8
      */
-    private void registerServiceBeans(Set<String> packagesToScan, BeanDefinitionRegistry registry) {
+    private void registerServiceAnnotationBeanPostProcessor(Set<String> packagesToScan, BeanDefinitionRegistry registry) {
 
-        DubboClassPathBeanDefinitionScanner dubboClassPathBeanDefinitionScanner =
-                new DubboClassPathBeanDefinitionScanner(registry, environment, resourceLoader);
+        BeanDefinitionBuilder builder = rootBeanDefinition(ServiceAnnotationBeanPostProcessor.class);
+        builder.addConstructorArgValue(packagesToScan);
+        builder.setRole(BeanDefinition.ROLE_INFRASTRUCTURE);
+        AbstractBeanDefinition beanDefinition = builder.getBeanDefinition();
+        BeanDefinitionReaderUtils.registerWithGeneratedName(beanDefinition, registry);
 
-        dubboClassPathBeanDefinitionScanner.addIncludeFilter(new AnnotationTypeFilter(Service.class));
-
-        for (String packageToScan : packagesToScan) {
-
-            Set<BeanDefinitionHolder> beanDefinitionHolders = dubboClassPathBeanDefinitionScanner.doScan(packageToScan);
-
-            for (BeanDefinitionHolder beanDefinitionHolder : beanDefinitionHolders) {
-                registerServiceBean(beanDefinitionHolder, registry);
-            }
-
-            if (logger.isInfoEnabled()) {
-                logger.info(beanDefinitionHolders.size() + " annotated @Service Components { " +
-                        beanDefinitionHolders +
-                        " } were scanned under package[" + packageToScan + "]");
-            }
-        }
-
-    }
-
-    private Class<?> resolveClass(BeanDefinitionHolder beanDefinitionHolder) {
-
-        BeanDefinition beanDefinition = beanDefinitionHolder.getBeanDefinition();
-
-        return resolveClass(beanDefinition);
-
-    }
-
-    private Class<?> resolveClass(BeanDefinition beanDefinition) {
-
-        String beanClassName = beanDefinition.getBeanClassName();
-
-        return resolveClassName(beanClassName, classLoader);
-
-    }
-
-    /**
-     * Registers {@link ServiceBean} from new annotated {@link Service} {@link BeanDefinition}
-     *
-     * @param beanDefinitionHolder
-     * @param registry
-     * @see ServiceBean
-     * @see BeanDefinition
-     */
-    private void registerServiceBean(BeanDefinitionHolder beanDefinitionHolder, BeanDefinitionRegistry registry) {
-
-        Class<?> beanClass = resolveClass(beanDefinitionHolder);
-
-        Service service = findAnnotation(beanClass, Service.class);
-
-        Class<?> interfaceClass = resolveServiceInterfaceClass(beanClass, service);
-
-        String beanName = beanDefinitionHolder.getBeanName();
-
-        AbstractBeanDefinition serviceBeanDefinition = buildServiceBeanDefinition(service, interfaceClass, beanName);
-
-        registerWithGeneratedName(serviceBeanDefinition, registry);
-
-    }
-
-    private ManagedList<RuntimeBeanReference> toRuntimeBeanReferences(String... beanNames) {
-
-        ManagedList<RuntimeBeanReference> runtimeBeanReferences = new ManagedList<RuntimeBeanReference>();
-
-        if (!ObjectUtils.isEmpty(beanNames)) {
-
-            for (String beanName : beanNames) {
-
-                String resolvedBeanName = environment.resolvePlaceholders(beanName);
-
-                runtimeBeanReferences.add(new RuntimeBeanReference(resolvedBeanName));
-            }
-
-        }
-
-        return runtimeBeanReferences;
-
-    }
-
-    private void addPropertyReference(BeanDefinitionBuilder builder, String propertyName, String beanName) {
-        String resolvedBeanName = environment.resolvePlaceholders(beanName);
-        builder.addPropertyReference(propertyName, resolvedBeanName);
-    }
-
-
-    private AbstractBeanDefinition buildServiceBeanDefinition(Service service, Class<?> interfaceClass,
-                                                              String annotatedServiceBeanName) {
-
-        BeanDefinitionBuilder builder = rootBeanDefinition(ServiceBean.class)
-                .addConstructorArgValue(service)
-                // References "ref" property to annotated-@Service Bean
-                .addPropertyReference("ref", annotatedServiceBeanName)
-                .addPropertyValue("interfaceClass", interfaceClass);
-
-        /**
-         * Add {@link com.alibaba.dubbo.config.ProviderConfig} Bean reference
-         */
-        String providerConfigBeanName = service.provider();
-        if (StringUtils.hasText(providerConfigBeanName)) {
-            addPropertyReference(builder, "provider", providerConfigBeanName);
-        }
-
-        /**
-         * Add {@link com.alibaba.dubbo.config.MonitorConfig} Bean reference
-         */
-        String monitorConfigBeanName = service.monitor();
-        if (StringUtils.hasText(monitorConfigBeanName)) {
-            addPropertyReference(builder, "monitor", monitorConfigBeanName);
-        }
-
-        /**
-         * Add {@link com.alibaba.dubbo.config.ApplicationConfig} Bean reference
-         */
-        String applicationConfigBeanName = service.application();
-        if (StringUtils.hasText(applicationConfigBeanName)) {
-            addPropertyReference(builder, "application", applicationConfigBeanName);
-        }
-
-        /**
-         * Add {@link com.alibaba.dubbo.config.ModuleConfig} Bean reference
-         */
-        String moduleConfigBeanName = service.module();
-        if (StringUtils.hasText(moduleConfigBeanName)) {
-            addPropertyReference(builder, "module", moduleConfigBeanName);
-        }
-
-
-        /**
-         * Add {@link com.alibaba.dubbo.config.RegistryConfig} Bean reference
-         */
-        String[] registryConfigBeanNames = service.registry();
-
-        List<RuntimeBeanReference> registryRuntimeBeanReferences = toRuntimeBeanReferences(registryConfigBeanNames);
-
-        if (!registryRuntimeBeanReferences.isEmpty()) {
-            builder.addPropertyValue("registries", registryRuntimeBeanReferences);
-        }
-
-        /**
-         * Add {@link com.alibaba.dubbo.config.ProtocolConfig} Bean reference
-         */
-        String[] protocolConfigBeanNames = service.protocol();
-
-        List<RuntimeBeanReference> protocolRuntimeBeanReferences = toRuntimeBeanReferences(protocolConfigBeanNames);
-
-        if (!registryRuntimeBeanReferences.isEmpty()) {
-            builder.addPropertyValue("protocols", protocolRuntimeBeanReferences);
-        }
-
-        return builder.getBeanDefinition();
-
-    }
-
-    private Class<?> resolveServiceInterfaceClass(Class<?> annotatedServiceBeanClass, Service service) {
-
-        Class<?> interfaceClass = service.interfaceClass();
-
-        if (void.class.equals(interfaceClass)) {
-
-            interfaceClass = null;
-
-            String interfaceClassName = service.interfaceName();
-
-            if (StringUtils.hasText(interfaceClassName)) {
-                if (ClassUtils.isPresent(interfaceClassName, classLoader)) {
-                    interfaceClass = resolveClassName(interfaceClassName, classLoader);
-                }
-            }
-
-        }
-
-        if (interfaceClass == null) {
-
-            Class<?>[] allInterfaces = annotatedServiceBeanClass.getInterfaces();
-
-            if (allInterfaces.length > 0) {
-                interfaceClass = allInterfaces[0];
-            }
-
-        }
-
-        Assert.notNull(interfaceClass,
-                "@Service interfaceClass() or interfaceName() or interface class must be present!");
-
-        Assert.isTrue(interfaceClass.isInterface(),
-                "The type that was annotated @Service is not an interface!");
-
-        return interfaceClass;
     }
 
     /**
@@ -295,21 +92,6 @@ public class DubboComponentScanRegistrar implements ImportBeanDefinitionRegistra
             return Collections.singleton(ClassUtils.getPackageName(metadata.getClassName()));
         }
         return packagesToScan;
-    }
-
-    @Override
-    public void setResourceLoader(ResourceLoader resourceLoader) {
-        this.resourceLoader = resourceLoader;
-    }
-
-    @Override
-    public void setEnvironment(Environment environment) {
-        this.environment = environment;
-    }
-
-    @Override
-    public void setBeanClassLoader(ClassLoader classLoader) {
-        this.classLoader = classLoader;
     }
 
 }
